+++ generators/app/templates/ts/src/components/app/footer/Footer.tsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import './Footer.scss';
 
 const Footer: React.FC = () => (