+++ generators/app/templates/ts/src/dev-client.d.ts
@@ -1,0 +1,5 @@
+/**
+ * Declares global module for webpack-hot-middleware dev-client.
+ * @packageDocumentation
+ */
+declare module 'webpack-hot-middleware/client';