+++ generators/app/templates/ts/src/index.tsx
@@ -1,6 +1,6 @@
-import * as React from 'react';
+import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
-import * as ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom';
 import App from './components/app';
 import './scss/style.scss';