+++ generators/app/templates/ts/src/typings.d.ts
@@ -1,5 +1,0 @@
-/**
- * Declares type roots.
- * @packageDocumentation
- */
-declare module '*.md';